FOGL-828 separated routes from service_registry
