+++ python/foglamp/services/core/connect.py
@@ -5,7 +5,7 @@
 # FOGLAMP_END
 
 
-from foglamp.services.common.microservice_management.service_registry.service_registry import Service
+from foglamp.services.core.service_registry.service_registry import Service
 from foglamp.common.storage_client.storage_client import StorageClient
 from foglamp.common import logger
