+++ python/foglamp/services/core/service_registry/service_registry.py
